+++ CHANGELOG.md
@@ -7,6 +7,39 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Sharing popover button to the discover search page
+- `scheduled-banner` component that pulls banners (created in the OSF Admin app) from the API.
+
+### Changed
+- Format last edited date in search result like "MMM DDD, YYYY UTC" instead of "YYYY-MM-DD (UTC)"
+- `navbar-auth-dropdown` to match current changes to the navbar
+
+### Fixed
+- Missing banner translation error
+
+## [0.16.2] - 2018-05-01
+### Added
+- temporary inline style to sign-up button
+
+## [0.16.1] - 2018-04-26
+### Added
+- `activeFilters.types` reset when registration provider changes
+
+## [0.16.0] - 2018-04-24
+### Added
+- `documentType` property to `preprint-service` model.
+- `secondary-nav-dropdown` class to `new-navbar-auth-dropdown` component
+- `ember-component-css` to bundle all component styles in one `pod-styles.scss` and prevent styling conflicts across components.
+- `works` to preprint word translations
+
+### Changed
+- static strings to be in translation files
+- `bare-strings` option in eslint to `true`
+- OSF API version from 2.4 to 2.6
+- ember-cli-moment-shim version to `^3.5.3` due to security issues found in `moment` versions before `2.19.3`
+- classes and element tags to match styles needed for long name truncation on navbar
+
+### Added
+- Add moderator model, serializer, and adapter
 
 ## [0.15.0] - 2018-03-06
 ### Added